requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
feedparser>=6.0.0
requests>=2.31.0
